pyproject.toml
[tool.poetry]
name = "mirage-defect"
version = "0.1.0"
description = ""
authors = ["Shunsuke NAKATSUKA <jacketgardensio@gmail.com>"]
readme = "README.md"
packages = [
  { include = "mirage_defect/" }
]

[tool.poetry.dependencies]
python = "^3.9"
torch = {version = "^2.0.0", source = "torch_cu118"}
torchvision = {version = "^0.15.0", source = "torch_cu118"}
torchaudio = {version = "^2.0.0", source = "torch_cu118"}
triton = {version = "^2.0.0", source = "torch_cu118"}
segment-anything = { git = "https://github.com/salty-vanilla/segment-anything" }
transformers = "^4.35.2"
addict = "^2.4.0"
yapf = "^0.40.2"
timm = "^0.9.11"
numpy = "^1.26.2"
opencv-python = "^4.8.1.78"
supervision = "0.6.0"
pycocotools = "^2.0.7"
groundingdino-py = "0.4.0"
pydantic = "^2.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "torch_cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
notebook = "^7.0.6"
ipywidgets = "^8.1.1"

[tool.poetry.extras]
all = ["torch"]
